Restructure translator module
